Rename draggable_sheet_extent
